* origin/released:
  xserver libX11 xcb-proto mesa pixman plink git update 2 Nov 2015

+++ xorg-server/glx/glxcmds.c
@@ -2187,7 +2187,7 @@ __glXDisp_RenderLarge(__GLXclientState * cl, GLbyte * pc)
         __GLXrenderSizeData entry;
         int extra = 0;
         int left = (req->length << 2) - sz_xGLXRenderLargeReq;
-        size_t cmdlen;
+        int cmdlen;
         int err;
 
         /*
